storage: move read-only table worker to own module
In preparation for some more refactorings, and largely to get it out of
the way in that bigger persist_handles.rs.
